feats: correção de bugs e implementações

+++ src/services/UserContext.jsx
@@ -2,10 +2,9 @@
 import React, { createContext, useContext, useState, useEffect } from 'react';
 import { useNavigate } from 'react-router-dom';
 // auth:
-import { GoogleAuthProvider, signInWithPopup, getAuth, signInWithEmailAndPassword, createUserWithEmailAndPassword, sendPasswordResetEmail, signOut, onAuthStateChanged, updateProfile, updatePassword, sendEmailVerification, updateEmail, deleteUser, reauthenticateWithCredential, EmailAuthProvider } from "firebase/auth";
+import { GoogleAuthProvider, signInWithPopup, signInWithEmailAndPassword, createUserWithEmailAndPassword, sendPasswordResetEmail, signOut, onAuthStateChanged, updateProfile, updateEmail, deleteUser, reauthenticateWithCredential, EmailAuthProvider } from "firebase/auth";
 // firestore: 
-import { addDoc, collection, deleteDoc, doc, getDocs, updateDoc, getDoc, query, where, arrayUnion, serverTimestamp, arrayRemove, increment } from "firebase/firestore";
-import { useCollectionData } from 'react-firebase-hooks/firestore';
+import { addDoc, collection, deleteDoc, doc, getDocs, updateDoc, query, where, increment, arrayUnion, arrayRemove, serverTimestamp, orderBy } from "firebase/firestore";
 //moment:
 import moment from 'moment/moment';
 
@@ -28,15 +27,12 @@ export const CostumerProvider = ({ children }) => {
     const [bios, setBios] = useState('');
     const [name, setName] = useState('');
     const [categorys, setCategorys] = useState('');
+    const [following, setFollowing] = useState([]);
+    const [followers, setFollowers] = useState([]);
+    const [authProviderUser, setAuthProviderUser] = useState('');
     const [uposts, setUposts] = useState([]);
     const [reviews, setReviews] = useState([]);
     const [allReviews, setAllReviews] = useState([]);
-    const [followers, setFollowers] = useState([]);
-    const [following, setFollowing] = useState([]);
-
-    // usuario externo/especifico 
-    const [euser, setEuser] = useState(null);
-    const [eposts, setEposts] = useState([]);
 
     // dados do feed
     const [users, setUsers] = useState([]);
@@ -52,7 +48,6 @@ export const CostumerProvider = ({ children }) => {
         const unsubscribe = onAuthStateChanged(auth, (currentUser) => {
             setUser(currentUser);
             setUid(currentUser.uid);
-            // console.log('user setado pelos dados do currentUser/ useEffecct Context', user);
             setName(currentUser.displayName);
             setImgUrl(currentUser.photoURL);
             getUserId();
@@ -82,6 +77,9 @@ export const CostumerProvider = ({ children }) => {
                 setImgUrl(doc.data().imgURL ? doc.data().imgURL : user.photoURL);
                 setBios(doc.data().userBio);
                 setCategorys(doc.data().userCategorys);
+                setAuthProviderUser(doc.data().authProvider);
+                setFollowing(doc.data().following ? doc.data().following : []);
+                setFollowers(doc.data().followers ? doc.data().followers : []);
 
                 if (name) {
                     break
@@ -96,7 +94,7 @@ export const CostumerProvider = ({ children }) => {
         setUposts(userPosts);
     }
 
-    // que pegue todos os users da coleção para rederiza-los no filtro de busca e para filtrar o usuário logado e externo  
+    // que pegue todos os users da coleção, para realizar filtragens e mais 
     const getUsers = async () => {
         const q = query(collection(db, "users"));
         const querySnapshot = await getDocs(q);
@@ -108,7 +106,9 @@ export const CostumerProvider = ({ children }) => {
                     euid: doc.data().uid,
                     ename: doc.data().name ? doc.data().name : null,
                     eavatar: doc.data().imgURL ? doc.data().imgURL : null,
-                    ebios: doc.data().userBio ? doc.data().userBio : null
+                    ebios: doc.data().userBio ? doc.data().userBio : null,
+                    efollowers: doc.data().followers ? doc.data().followers : [],
+                    efollowing: doc.data().following ? doc.data().following : []
                 }
                 d.push(userObj);
             });
@@ -132,7 +132,9 @@ export const CostumerProvider = ({ children }) => {
                 imgURL: user.photoURL,
                 password,
                 authProvider: "local",
-                email
+                email,
+                following: [],
+                followers: []
             });
             setName(name);
             alert('usuario cadastrado no banco!');
@@ -222,7 +224,7 @@ export const CostumerProvider = ({ children }) => {
 
         // Create a reference under which you want to list
         const postsRef = ref(storage, `/postContent/user:${uid}`);
-        const profileRef = ref(storage, `/profile/user:${uid}`)
+        const profileRef = ref(storage, `/profile/user:${uid}`);
 
         // Find all the prefixes and items.
         listAll(postsRef)
@@ -289,7 +291,9 @@ export const CostumerProvider = ({ children }) => {
                         name: user.displayName,
                         authProvider: "google",
                         email: user.email,
-                        imgURL: user.photoURL
+                        imgURL: user.photoURL,
+                        following: [],
+                        followers: []
                     });
                 }
 
@@ -369,39 +373,9 @@ export const CostumerProvider = ({ children }) => {
         } else {
             alert('E-mail não cadastrado! Cadastre-se no site ou coloque um e-mail cadastrado');
         }
-
-        // ------------
-
-
-        // if(validEmail){
-        // sendPasswordResetEmail(auth, email)
-        //     .then(() => {
-        //         alert('E-mail de redefinição de senha enviado');
-        //         navigate('/login');
-        //     })
-        //     .catch((error) => {
-        //         console.log(error);
-        //         alert(error.message, error.code);
-        //     });
-        // }else {
-        //     alert('E-mail não cadastrado! Cadastre-se no site ou coloque um e-mail cadastrado');
-        // }
     }
 
     const updateUserPasswordLogIn = async (email) => {
-        //usar status de email verificado ou não como condição para a query, se for verificado não há necessidade dela ser feita, caso contrário ela é feita para verificar se há um e-mail relacionado com o usuário logado no banco
-
-        // if (auth.currentUser.emailVerified) {
-        //     sendPasswordResetEmail(auth, auth.currentUser.email)
-        //         .then(() => {
-        //             alert('E-mail de redefinição de senha enviado/ e-mail autenticado!');
-        //             navigate('/config');
-        //         })
-        //         .catch((error) => {
-        //             console.log(error);
-        //             alert(error.message, error.code);
-        //         });
-        // } else {
         const q = query(collection(db, "users"), where('uid', '==', uid), where("email", "==", email));
         const querySnapshot = await getDocs(q);
         console.log(querySnapshot);
@@ -422,12 +396,9 @@ export const CostumerProvider = ({ children }) => {
         } else {
             alert('Não encontramos nenhum e-mail relacionado à sua conta');
         }
-        // }
-
     }
 
     const updateUserEmail = async (email, password) => {
-
         const credential = EmailAuthProvider.credential(
             auth.currentUser.email,
             password
@@ -441,17 +412,129 @@ export const CostumerProvider = ({ children }) => {
             console.log(error);
             alert(error.message);
         });
+    }
 
-        // getUserId();
+    //função para atualizar numero de seguidores e seguindo
+    const addFollowing = async (euid) => {
+        //atualizando doc do usuário logado adicionando em um array (following) o uid que que ele quer seguir
+        try {
+            await updateDoc(doc(db, "users", id), {
+                following: arrayUnion(euid)
+            });
+        } catch (error) {
+            console.log(error);
+        }
 
-        // sendEmailVerification(auth.currentUser)
-        //     .then(() => {
-        //         alert('E-mail de verificação enviado! ');
-        //     })
-        //     .catch((error) => {
-        //         console.log(error);
-        //         alert(error.message);
-        //     })
+        //para adicionar o uid do usuário logado aos dados do user a quem ele está seguindo (o outro ganha um seguidor)
+        const q = query(collection(db, "users"), where("uid", "==", euid));
+        const querySnapshot = await getDocs(q);
+        let eid = '';
+        querySnapshot.forEach((doc) => {
+            eid = doc.id
+        });
+        try {
+            await updateDoc(doc(db, "users", eid), {
+                followers: arrayUnion(uid)
+            });
+        } catch (error) {
+            console.log(error)
+        }
+
+    }
+
+    const removeFollowing = async (euid) => {
+        try {
+            await updateDoc(doc(db, "users", id), {
+                following: arrayRemove(euid)
+            });
+            alert('adicionando na lista de seguindo');
+        } catch (error) {
+            console.log(error);
+            alert(error.message);
+        }
+
+        //para remover o uid do usuário logado dos dados do user a quem ele está seguindo
+        const q = query(collection(db, "users"), where("uid", "==", euid));
+        const querySnapshot = await getDocs(q);
+        let eid = '';
+        querySnapshot.forEach((doc) => {
+            eid = doc.id
+        });
+        try {
+            await updateDoc(doc(db, "users", eid), {
+                followers: arrayRemove(uid)
+            });
+        } catch (error) {
+            console.log(error);
+            alert(error.message);
+        }
+
+    }
+
+    //funções dos comentários
+    const addComent = async (idPost, content, imgContent) => {
+        try {
+            await addDoc(collection(db, "post"), {
+                uid: uid,
+                userPhoto: user?.photoURL ? user?.photoURL : null,
+                name: name,
+                content: content,
+                img_content: imgContent ? imgContent : null,
+                postFather: idPost,
+                date: serverTimestamp()
+            });
+            await updateDoc(doc(db, 'post', idPost), {
+                coments: increment(1)
+            });
+            getPosts();
+
+        } catch (error) {
+            console.log(error);
+            alert(error.message);
+        }
+    }
+
+    const updateComent = (comentId, content, imgContent) => {
+        updateDoc(doc(db, 'post', comentId), {
+            content: content,
+            imgContent: imgContent
+        }).then(() => {
+            alert('comentáio atualizado');
+            getPosts();
+        }).catch((error) => {
+            console.log(error);
+            alert(error.message);
+        });
+    }
+    const removeComent = (comentId, idPost, itemRef) => {
+        deleteDoc(doc(db, 'post', comentId))
+            .then(() => {
+                alert('comentário deletado');
+            })
+            .catch((error) => {
+                console.log(error);
+                alert(error.message);
+            });
+
+        updateDoc(doc(db, 'post', idPost), {
+            coments: increment(-1)
+        }).then(() => {
+            alert('comentarios de post atualizado');
+            getPosts();
+        }).catch((error) => {
+            console.log(error);
+            alert(error.message);
+        });
+
+        if (itemRef !== null) {
+            const imgRef = ref(storage, itemRef);
+            deleteObject(imgRef).then(() => {
+                // File deleted successfully
+                console.log('reference deletada hi from imgReference: ');
+            }).catch((error) => {
+                console.log(error);
+            });
+        }
     }
 
     /* ================================
@@ -463,13 +546,15 @@ export const CostumerProvider = ({ children }) => {
         try {
             const createdPost = await addDoc(collection(db, "post"), {
                 uid: user?.uid,
-                userPhoto: user?.photoURL ? user?.photoURL : null,
+                userPhoto: imgUrl ? imgUrl : null,
                 name: name,
                 title: title,
                 content: content,
                 category: cat,
                 imgContent: img_content ? img_content : null,
-                likes: 0
+                likes: 0,
+                coments: 0,
+                date: serverTimestamp()
             });
 
             registerReview(
@@ -488,18 +573,18 @@ export const CostumerProvider = ({ children }) => {
             getPosts();
             getUserId();
             getReviews();
-            alert('Flashcard e revisão criados!');
+            alert('Flashcard criado!');
         } catch (err) {
-            console.log(err)
+            console.log(err);
+            alert(err.message);
         }
 
     }
 
     // que pegue todos os posts da tabela para renderizar no explore 
     const getPosts = async () => {
-
         try {
-            const querySnapshot = await getDocs(collection(db, "post"));
+            const querySnapshot = await getDocs(collection(db, "post"), orderBy("date", "desc"));
             const d = []
             querySnapshot.forEach((doc) => {
                 // doc.data() is never undefined for query doc snapshots
@@ -512,7 +597,9 @@ export const CostumerProvider = ({ children }) => {
                     category: doc.data().category,
                     content: doc.data().content,
                     img_content: doc.data().imgContent,
-                    likes: doc.data().likes
+                    likes: doc.data().likes,
+                    coments: doc.data().coments,
+                    postFather: doc.data().postFather
                 }
                 d.push(post);
             });
@@ -532,13 +619,6 @@ export const CostumerProvider = ({ children }) => {
             getPosts();
         });
 
-        // const aq = query(collection(db, "post"), where('postId', '==', postId));
-        // const aquerySnapshot = await getDocs(q);
-        // let uimgRef = ''
-        // aquerySnapshot.forEach((doc) => {
-        //     uimgRef = doc.data().imgContent
-        // });
-
         if (itemRef !== null) {
             const imgRef = ref(storage, itemRef);
             deleteObject(imgRef).then(() => {
@@ -554,14 +634,10 @@ export const CostumerProvider = ({ children }) => {
         querySnapshot.forEach((doc) => {
             deleteDoc(doc(db, 'revision', doc.id));
         });
-
-        getReviews();
-        getUserId();
     }
 
     //FUNÇÃO DE ATUALIZAR POST
     const updatePost = async (postId, title, cat, content, img_content) => {
-
         updateDoc(doc(db, 'post', postId), {
             title: title,
             category: cat,
@@ -569,8 +645,6 @@ export const CostumerProvider = ({ children }) => {
             imgContent: img_content,
         }).then(() => {
             alert('Mudança no post feita');
-            getPosts();
-            getUserId();
         }).catch((error) => {
             console.log(error);
             alert(error.message);
@@ -589,13 +663,13 @@ export const CostumerProvider = ({ children }) => {
             imgContent: img_content
         }).then(() => {
             alert('Mudança na revisão feita');
-            getPosts();
-            getUserId();
         }).catch((error) => {
             console.log(error);
             alert(error.message);
         });
 
+        getPosts();
+        getUserId();
     }
 
     //para adicionar adicionar 1 like no post
@@ -653,27 +727,37 @@ export const CostumerProvider = ({ children }) => {
     }
 
     //para deletar revisão
-    const removeReview = async (postId) => {
-        try {
-            const q = query(collection(db, "revision"), where('postId', '==', postId));
-            const querySnapshot = await getDocs(q);
-            let eid = ''
-            querySnapshot.forEach((doc) => {
-                eid = doc.id
+    const removeReview = async (postId, reviewId) => {
+        if (reviewId) {
+            deleteDoc(doc(db, 'revision', reviewId)).then(() => {
+                alert('revisão deletada');
+            }).catch((error) => {
+                console.log(error);
+                alert(error.message);
             })
-            deleteDoc(doc(db, 'revision', eid));
-            getReviews();
-            getPosts();
-            alert('revisão deletada');
-        } catch (error) {
-            console.log(error);
-            alert(error.message);
+        }
+
+        if (!reviewId) {
+            try {
+                const q = query(collection(db, "revision"), where('postId', '==', postId), where('userAdded', '==', uid));
+                const querySnapshot = await getDocs(q);
+                let eid = ''
+                querySnapshot.forEach((doc) => {
+                    eid = doc.id
+                })
+                deleteDoc(doc(db, 'revision', eid));
+                getReviews();
+                getPosts();
+                alert('revisão deletada');
+            } catch (error) {
+                console.log(error);
+                alert(error.message);
+            }
         }
     }
 
     // que pegue todos as reviews da tabela para renderizar na rota review de acordo com sua data certa
     const getReviews = async () => {
-
         // pegue todas as revisoes do usuário logado
         const q = query(collection(db, "revision"), where('userAdded', '==', uid));
         const querySnapshot = await getDocs(q);
@@ -733,10 +817,9 @@ export const CostumerProvider = ({ children }) => {
             allData.push(review);
             setAllReviews(allData);
         })
-
     }
 
-    //TODO -> FUNÇÃO PARA ATUALIZAR DATA DE REVISÃO
+    //FUNÇÃO PARA ATUALIZAR DATA DE REVISÃO
     const updateReview = async (reviewId, newDate, counter) => {
         // ao clicar no btn já revisei a revisao é atualizada para uma nova data de revisao e um novo contador
         try {
@@ -754,19 +837,16 @@ export const CostumerProvider = ({ children }) => {
     //INSTANCIA COSTUMER CONTEXT SENDO RETORNADA NO COMPONENTE PASSANDO PARA O SEU PROVEDOR AS FUNÇÕES CRUD, LOGIN E LOGOUT
     return (
         <CostumerContext.Provider value={{
-            user, uid, id, imgUrl, bios, categorys, uposts, followers, following, euser, users,
-            eposts, posts,
+            user, uid, id, imgUrl, bios, categorys, uposts, followers, following, authProviderUser,
+            users, posts,
             reviews, notification, allReviews,
 
             registerWithEmailAndPassword, logInWithEmailAndPassword, signInWithGoogle, logout,
             updateUserPasswordLogOut, updateUserPasswordLogIn, updateUserEmail,
 
-            getUsers, getUserId, updateUserProfile, revomeUser,
-            // getExternalUser, 
-            // addFollowing, removeFollowing,
+            getUsers, getUserId, updateUserProfile, revomeUser, addFollowing, removeFollowing,
 
-            getPosts, registerPost, updatePost, deletePost, addLikePost, removeLikePost,
-            // getExternalPost, 
+            getPosts, registerPost, updatePost, deletePost, addLikePost, removeLikePost, addComent, removeComent, updateComent,
 
             registerReview, getReviews, updateReview, removeReview
         }}>
